requirements.txt
streamlit==1.56.0
python-dotenv==1.2.2
langchain==1.2.15
langchain-core==1.2.26
langchain-openai==1.1.12
openai==2.30.0
Flask==2.3.3
